+++ spec/connectors/importer_spec.rb
@@ -715,6 +715,27 @@ describe CartoDB::Connector::Importer do
       @user.save
     end
 
+    it 'fails to import a visualization export if public dataset quota is exceeded' do
+      filepath = "#{Rails.root}/services/importer/spec/fixtures/public_visualization_export_with_csv_table.carto"
+      @user.public_dataset_quota = 0
+      @user.save
+      @data_import = DataImport.create(
+        user_id: @user.id,
+        data_source: filepath,
+        updated_at: Time.now.utc,
+        append: false,
+        create_visualization: true
+      )
+      @data_import.values[:data_source] = filepath
+
+      expect { @data_import.run_import! }.to raise_error('Public dataset quota exceeded')
+      @data_import.success.should eq false
+      @data_import.error_code.should eq 8008
+
+      @user.public_dataset_quota = nil
+      @user.save
+    end
+
     it 'fails to import a visualization export if private map quota is exceeded' do
       filepath = "#{Rails.root}/services/importer/spec/fixtures/visualization_export_with_csv_table.carto"
       @user.private_map_quota = 0